Solved day 3

+++ src/day01.rs
@@ -1,7 +1,7 @@
 use std::collections::HashSet;
 
-pub fn part1(input: String) -> i32 {
-    let nums: HashSet<i32> = input.lines().map(|x| x.parse().unwrap()).collect();
+pub fn part1(input: String) -> usize {
+    let nums: HashSet<usize> = input.lines().map(|x| x.parse().unwrap()).collect();
     for &x in nums.iter() {
         let y = 2020 - x;
         if nums.contains(&y) {
@@ -11,8 +11,8 @@ pub fn part1(input: String) -> i32 {
     0
 }
 
-pub fn part2(input: String) -> i32 {
-    let nums: HashSet<i32> = input.lines().map(|x| x.parse().unwrap()).collect();
+pub fn part2(input: String) -> usize {
+    let nums: HashSet<usize> = input.lines().map(|x| x.parse().unwrap()).collect();
     for &x in nums.iter() {
         for &y in nums.iter() {
             if x + y > 2020 {
